更新

+++ lib/flutter_hpplay.dart
@@ -19,8 +19,8 @@ class FlutterHpplay {
   List serviceNames = [];
   int currentIndex = -1;
   LBLelinkPlayStatus playStatus;
-  double duration;
-  double currentTime;
+  int duration;
+  int currentTime;
 
   DynamicHandler _onLelinkBrowserError;
   DynamicHandler _onLelinkBrowserDidFindLelinkServices;
@@ -133,8 +133,8 @@ class FlutterHpplay {
       },
       onLelinkPlayerProgressInfo: (Map<String, dynamic> message) async {
         _onLelinkPlayerProgressInfo(message);
-        duration = double.parse(message['duration']);
-        currentTime = double.parse(message['currentTime']);
+        duration = int.parse(message['duration']);
+        currentTime = int.parse(message['currentTime']);
         // print(
         //     'flutter--播放进度 总时长:${message['duration']}、当前播放位置:${message['currentTime']}');
       },
